Split stream module out of feed to separate responsibility for capturing frames and displaying frames. This lets me use the main process to manage displaying frames, while letting subprocesses determine what frames are displayed.
